feat: 用户注销功能

+++ backend/src/routes/UserRouter.ts
@@ -198,7 +198,18 @@ userRouter.post(
   '/delete',
   asyncWrapper(async (req: any, res, next) => {
     try {
-      res.status(200).json(await Service.Delete(req.auth.id))
+      if (req.auth.group === Group.SUPER_ADMIN) {
+        res
+          .status(200)
+          .json(
+            new Restful(
+              CodeDictionary.DELETE_ERROR__USER_ADMIN,
+              '不能删除超级管理员账号',
+            ),
+          )
+      } else {
+        res.status(200).json(await Service.Delete(req.auth.id))
+      }
     } catch (e) {
       // TODO: 进行邮件提醒
       res.status(500).end()
